fixes doc for deepMapValues and makes consistent with lodash docs
the second argument is not 'propertyPath', it is the 'iteratee' function. Below, this it is referred to as 'callback', which is a little vague. lodash docs always refer to this function as 'iteratee' which say a lot about what the purpose of the function is. I've updated the documentation to make it consistent with lodash docs.